init commit

+++ database/migrations/2024_10_23_214000_create_tickets_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('name');
             $table->string('email')->nullable();
-            $table->integer('number')->nullable();
+            $table->bigInteger('number')->nullable();
             $table->text('description');
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('ticket_id')->nullable()->constrained('tickets', 'id')->cascadeOnDelete();
